+++ context/app/api/vitessce_confs/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa: F401
-from .assay_confs import get_view_config_builder
+from .builder_factory import get_view_config_builder
